Start converting to Swift
Bit of a bulk commit here, but I forgot to do nice small commits early on…and well, here we are.

Right now the actual existing functionality hasn’t all been completely proved over. The hue-dansen.c executable needs to be ported still, but I *think* the functionality in hue-dtls.c has been properly ported.

One of the goals of this is to write cross-platform code, so there are some areas that I’m intentionally steering clear of. For example, Networking.framework looks like a nicer API than futzing around with OpenSSL, but it’s only available on Apple platforms. Eventually I’d like to switch to using SwiftNIO, but the it doesn’t currently support DTLS (might be an interesting area to add in later though).